refactor : study list color set
