+++ Terminal-iOS/Study/StudyCategory/CategoryCell.swift
@@ -36,7 +36,7 @@ class CategoryCell: UICollectionViewCell {
     }
     
     private func layout() {
-        contentView.addSubview(imageView)
+//        contentView.addSubview(imageView)
         addSubview(imageView)
         imageView.translatesAutoresizingMaskIntoConstraints = false
         imageView.topAnchor.constraint(equalTo: topAnchor).isActive = true
