+++ Terminal-iOS/Study/StudyCategory/StudyCategoryView.swift
@@ -35,9 +35,6 @@ class StudyCategoryView: UIViewController {
         let searchStudyBtn = UIBarButtonItem(barButtonSystemItem: .search,
                                              target: self,
                                              action: #selector(searchStudy))
-        view.do {
-            $0.backgroundColor = UIColor(named: "background")
-        }
         self.do {
             $0.view.backgroundColor = UIColor.appColor(.terminalBackground)
             $0.title = "스터디"
